+++ SimGeneral/PreMixingModule/BuildFile.xml
@@ -1,0 +1,7 @@
+<use   name="FWCore/Framework"/>
+<use   name="FWCore/ParameterSet"/>
+<use   name="FWCore/PluginManager"/>
+<use   name="FWCore/ServiceRegistry"/>
+<export>
+  <lib   name="1"/>
+</export>
