+++ SimGeneral/PreMixingModule/plugins/BuildFile.xml
@@ -1,27 +1,14 @@
-<use   name="CondFormats/SiPixelObjects"/>
-<use   name="CondFormats/SiStripObjects"/>
 <use   name="DataFormats/HepMCCandidate"/>
-<use   name="DataFormats/TrackerCommon"/>
-<use   name="DataFormats/EcalDigi"/>
-<use   name="DataFormats/HcalDigi"/>
-<use   name="DataFormats/DTDigi"/>
-<use   name="DataFormats/RPCDigi"/>
-<use   name="DataFormats/CSCDigi"/>
 <use   name="FWCore/Framework"/>
 <use   name="FWCore/MessageLogger"/>
 <use   name="FWCore/ParameterSet"/>
 <use   name="FWCore/PluginManager"/>
 <use   name="FWCore/ServiceRegistry"/>
-<use   name="Geometry/TrackerGeometryBuilder"/>
-<use   name="Geometry/Records"/>
 <use   name="Mixing/Base"/>
 <use   name="SimDataFormats/CrossingFrame"/>
 <use   name="SimDataFormats/PileupSummaryInfo"/>
-<use   name="SimTracker/SiStripDigitizer"/>
-<use   name="SimCalorimetry/EcalSimProducers"/>
-<use   name="SimCalorimetry/EcalSimAlgos"/>
-<use   name="SimCalorimetry/HcalSimProducers"/>
-<use   name="SimCalorimetry/HcalSimAlgos"/>
+<use   name="SimGeneral/MixingModule"/>
+<use   name="SimGeneral/PreMixingModule"/>
 <library   file="*.cc" name="SimGeneralPreMixingModulePlugins">
   <flags   EDM_PLUGIN="1"/>
 </library>
