+++ SimGeneral/PreMixingModule/src/PreMixingWorkerFactory.cc
@@ -1,3 +1,3 @@
-#include "PreMixingWorkerFactory.h"
+#include "SimGeneral/PreMixingModule/interface/PreMixingWorkerFactory.h"
 
 EDM_REGISTER_PLUGINFACTORY(PreMixingWorkerFactory, "PreMixingWorkerFactory");
