+++ SimGeneral/TrackingAnalysis/plugins/BuildFile.xml
@@ -1,7 +1,9 @@
 <use   name="SimGeneral/TrackingAnalysis"/>
 <use   name="DataFormats/TrackerCommon"/>
 <use   name="DataFormats/HepMCCandidate"/>
+<use   name="SimDataFormats/TrackerDigiSimLink"/>
 <use   name="SimGeneral/MixingModule"/>
+<use   name="SimGeneral/PreMixingModule"/>
 <library   file="*.cc" name="SimGeneralTrackingAnalysisPlugins">
   <flags   EDM_PLUGIN="1"/>
 </library>
