+++ SimMuon/CSCDigitizer/plugins/BuildFile.xml
@@ -20,6 +20,7 @@
   <use   name="SimDataFormats/TrackerDigiSimLink"/>
   <use   name="SimGeneral/HepPDTRecord"/>
   <use   name="SimGeneral/NoiseGenerators"/>
+  <use   name="SimGeneral/PreMixingModule"/>
   <use   name="SimMuon/Neutron"/>
   <use   name="Utilities/General"/>
   <use   name="boost"/>
