+++ SimTracker/SiStripDigitizer/plugins/BuildFile.xml
@@ -1,4 +1,5 @@
 <library   file="*.cc" name="SimTrackerSiStripDigitizerPlugins">
   <use   name="SimTracker/SiStripDigitizer"/>
+  <use   name="SimGeneral/PreMixingModule"/>
   <flags   EDM_PLUGIN="1"/>
 </library>
